Use floats for percentages and add pod command

+++ internal/metrics/metrics.go
@@ -21,8 +21,8 @@ var (
 
 type MetricsValues struct {
 	Name       string
-	CPUPercent int
-	MemPercent int
+	CPUPercent float64
+	MemPercent float64
 	CPUCores   int
 	MemCores   int
 }
